requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.27
langchain-groq==0.1.9
langchain-huggingface
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.0.9
pypdf==4.3.1
pytesseract==0.3.10
scikit-learn==1.3.1
spacy==3.7.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==2.3.2
groq==0.9.0
Pillow==9.5.0
gunicorn
python-dotenv